contact_line: Improve detection of contact line
The contact line is now determined by rolling an atom across the
interface. This is more reliable since it avoid a slight fluctuation
along the y axis to sometimes become an advancement due to the binning
artifact.

Contact line advancement has been reworked. Now, the contact line
having advanced is determined in the following steps:

1) The (x,y) positions of the rolled contact line is checked and if Dx
is larger than the input dx the closest atom index is added as a
candidate for having advanced the contact line.

2) All candidates are checked to assert that they made a jump > hmin (a
new input value). This is to remove small fluctuations along y: the
contact line is only said to have advanced if the advancing molecule
made a proper jump (or is new).

3) MKT movement is determined as before.